+++ mods/BrewLAN/units/SEB5210/SEB5210_script.lua
@@ -4,21 +4,8 @@
 local TWallStructureUnit = import('/lua/terranunits.lua').TWallStructureUnit
 local CardinalWallUnit = import(import( '/lua/game.lua' ).BrewLANPath() .. '/lua/walls.lua').CardinalWallUnit
 --------------------------------------------------------------------------------
-local BrewLANPath = import( '/lua/game.lua' ).BrewLANPath()
-local OffsetBoneToTerrain = import(BrewLANPath .. '/lua/TerrainUtils.lua').OffsetBoneToTerrain
---------------------------------------------------------------------------------
 TWallStructureUnit = CardinalWallUnit( TWallStructureUnit )
 
-SEB5210 = Class(TWallStructureUnit) {
-    OnStopBeingBuilt = function(self,builder,layer)
-        TWallStructureUnit.OnStopBeingBuilt(self,builder,layer)
-        OffsetBoneToTerrain(self,'North')
-        OffsetBoneToTerrain(self,'South')
-        OffsetBoneToTerrain(self,'West')
-        OffsetBoneToTerrain(self,'East')
-        LOG("FEET")
-        LOG(repr(self.TerrainSlope))
-    end,
-}
+SEB5210 = Class(TWallStructureUnit) {}
 
 TypeClass = SEB5210